+++ src/body.h
@@ -2,16 +2,16 @@
 #define BODY_H
 
 typedef struct Vec2 {
-	float x, y;
+	double x, y;
 } Vec2;
 
 typedef struct Body {
-	float x, y;
-	float vx, vy;
-	float m;
+	double x, y;
+	double vx, vy;
+	double m;
 } Body;
 
-void update_vel(Body *bodies, size_t len, float G);
-void update_coords(Body *bodies, size_t len, float G);
+void update_vel(Body *bodies, size_t len, double G);
+void update_coords(Body *bodies, size_t len, double G);
 
 #endif
